requirements.txt
Flask==2.3.3
gradio==3.50.2
Pillow==10.0.1
Werkzeug==2.3.7
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3